# public/swagger.yaml
openapi: 3.0.0
info:
  title: O2 Concentrator Stockpile
  description: 'A stockpile service for O2 concentrators'
  termsOfService: http://swagger.io/terms/
  contact:
    email: adgang@tuta.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.1.0
...
servers:
- url: /{basePath}
  variables:
    basePath:
      default: api
- url: /{basePath}
  variables:
    basePath:
      default: api
...
paths:
...
  /donations:
    post:
      tags:
      - donations
      summary: Adds a donation
      description: Returns a single pet
      operationId: addDonation
      requestBody:
        description: Donation details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Donation'
        required: true
      responses:
        200:
          description: Donation added
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Donation'
            application/json:
              schema:
                $ref: '#/components/schemas/Donation'
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Donation not found
          content: {}
        405:
          description: Validation exception
          content: {}
      x-codegen-request-body-name: body
...
components:
  schemas:
...
    Address:
      type: object
      required:
      - line_1
      - line_2
      - area
      - city
      - district 
      - state 
      - pincode
      properties:
        line_1:
          type: string
        line_2:
          type: string
        area:
          type: string
        city:
          type: string
        district:
          type: string
        state:
          type: string
        pincode:
          type: string
...
    Donation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        payment_id:
          type: string
        name:
          type: string
        tax_id:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        email:
          type: string
        phone:
          type: string
...